+++ eng/external-plugin-validation.mjs
@@ -1,6 +1,7 @@
 import fs from "fs";
 import path from "path";
 import { ROOT_FOLDER } from "./constants.mjs";
+import { validateLicenseField } from "./lib/license.mjs";
 
 export const EXTERNAL_PLUGINS_FILE = path.join(ROOT_FOLDER, "plugins", "external.json");
 
@@ -48,45 +49,6 @@ const ALLOWED_SOURCE_KEYS = Object.freeze(["source", "repo", "path", "ref", "sha
 // with optional -prerelease and +build metadata.
 const SEMVER_PATTERN =
   /^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)(?:-((?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\.(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\+([0-9a-zA-Z-]+(?:\.[0-9a-zA-Z-]+)*))?$/;
-
-// A single SPDX license identifier token (e.g. "MIT", "Apache-2.0", "LicenseRef-Foo").
-const SPDX_ID_PATTERN = /^[A-Za-z0-9][A-Za-z0-9.-]*\+?$/;
-const SPDX_EXPRESSION_OPERATORS = new Set(["AND", "OR", "WITH"]);
-
-// Curated set of common SPDX license identifiers. Not exhaustive: unrecognized
-// but well-formed identifiers produce a warning rather than an error.
-const KNOWN_SPDX_IDS = new Set([
-  "0BSD",
-  "AGPL-3.0",
-  "AGPL-3.0-only",
-  "AGPL-3.0-or-later",
-  "Apache-2.0",
-  "BSD-2-Clause",
-  "BSD-3-Clause",
-  "BSL-1.0",
-  "CC-BY-4.0",
-  "CC-BY-SA-4.0",
-  "CC0-1.0",
-  "EPL-2.0",
-  "GPL-2.0",
-  "GPL-2.0-only",
-  "GPL-2.0-or-later",
-  "GPL-3.0",
-  "GPL-3.0-only",
-  "GPL-3.0-or-later",
-  "ISC",
-  "LGPL-2.1",
-  "LGPL-2.1-only",
-  "LGPL-2.1-or-later",
-  "LGPL-3.0",
-  "LGPL-3.0-only",
-  "LGPL-3.0-or-later",
-  "MIT",
-  "MPL-2.0",
-  "Unlicense",
-  "WTFPL",
-  "Zlib",
-]);
 
 // NOTE: Keep in sync with PLUGIN_JSON_CANDIDATES in external-plugin-quality-gates.mjs
 const EXTERNAL_PLUGIN_ROOT_MANIFEST_PATHS = Object.freeze([
@@ -265,85 +227,6 @@ function validateEmail(value, fieldName, prefix, errors) {
   if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(value)) {
     errors.push(`${prefix}: "${fieldName}" must be a valid email address`);
   }
-}
-
-function isRecognizedSpdxIdToken(token) {
-  if (token.startsWith("LicenseRef-") || token.startsWith("DocumentRef-")) {
-    return true;
-  }
-
-  if (!SPDX_ID_PATTERN.test(token)) {
-    return false;
-  }
-
-  const normalized = token.endsWith("+") ? token.slice(0, -1) : token;
-  return KNOWN_SPDX_IDS.has(normalized);
-}
-
-// Returns true only for a well-formed SPDX license expression composed of
-// recognized identifiers joined by AND/OR/WITH operators (optionally parenthesized).
-// Anything else (proprietary strings, free text, unrecognized ids) returns false so
-// the caller can warn without rejecting it — the plugin spec does not enforce SPDX.
-function isRecognizedSpdxExpression(license) {
-  const tokens = license
-    .replace(/[()]/g, " ")
-    .trim()
-    .split(/\s+/)
-    .filter((token) => token.length > 0);
-
-  if (tokens.length === 0) {
-    return false;
-  }
-
-  let expectOperator = false;
-  for (const token of tokens) {
-    if (SPDX_EXPRESSION_OPERATORS.has(token.toUpperCase())) {
-      if (!expectOperator) {
-        return false;
-      }
-      expectOperator = false;
-      continue;
-    }
-
-    if (expectOperator || !isRecognizedSpdxIdToken(token)) {
-      return false;
-    }
-
-    expectOperator = true;
-  }
-
-  return expectOperator;
-}
-
-// Canonical license validation shared by external plugins and local plugin.json
-// manifests. A non-SPDX license is a warning (never an error) so authors may use
-// proprietary or non-OSS licenses, per the agent-plugins-spec schema which does not
-// enforce SPDX. Returns collected errors/warnings.
-export function validateLicenseField(license, options = {}) {
-  const { required = false } = options;
-  const prefix = options.prefix ? `${options.prefix}: ` : "";
-  const errors = [];
-  const warnings = [];
-
-  if (license === undefined) {
-    if (required) {
-      errors.push(`${prefix}"license" is required`);
-    }
-    return { errors, warnings };
-  }
-
-  if (!isNonEmptyString(license)) {
-    errors.push(`${prefix}"license" must be a non-empty string`);
-    return { errors, warnings };
-  }
-
-  if (!isRecognizedSpdxExpression(license)) {
-    warnings.push(
-      `${prefix}"license" value "${license}" is not a recognized SPDX identifier; prefer a standard SPDX id (https://spdx.org/licenses), though non-SPDX or proprietary licenses are allowed`
-    );
-  }
-
-  return { errors, warnings };
 }
 
 function validateLicense(license, prefix, errors, warnings, required) {
